requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Cerberus==1.3.4
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
psycopg2==2.9.1
pycodestyle==2.8.0
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
